Add security for add, edit and delete products in the product app

+++ products/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, get_object_or_404, redirect, reverse
 from .models import Product, Category
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from .forms import ProductForm
 
 def all_products(request):
@@ -54,16 +55,19 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
- 
+@login_required 
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Request denied! You do not have authorisation to add products.')
+        return redirect(reverse('home'))
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
             individual_total
             form.save()
             messages.success(request, 'Successfully added product!')
-            return redirect(reverse('add_product'))
+            return redirect(reverse('product_detail', args=[product.id]))
         else:
             messages.error(request, 'Failed to add product. Please ensure the form is valid.')
     else:
@@ -75,8 +79,13 @@ def add_product(request):
  
     return render(request, 'products/add_product.html', context)
 
+
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Request denied! You do not have authorisation to edit products.')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -97,9 +106,12 @@ def edit_product(request, product_id):
 
     return render(request, 'products/edit_product.html', context)
 
-
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the site """
+    if not request.user.is_superuser:
+        messages.error(request, 'Request denied! You do not have authorisation to delete products.')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
